Add more functions to NamespaceTable struct.
